Add files via upload
In this assignment you are going to implement from scratch two cryptographic methods:
(i) Diffie-Hellman Key Exchange and (ii) RSA (Rivest–Shamir–Adleman). Both
implementations will be in the C programming language. The purpose of this
assignment is to provide you the opportunity to get familiar with the internals and
implementations of two popular encryption schemes, namely RSA and DH Key
Exchange. You will use the GMP C library to implement the RSA algorithm.